refactor(api): move revshare route to own file; extract `app`
