+++ api/src/index.ts
@@ -1,12 +1,7 @@
-import { Hono } from 'hono'
-import { cors } from 'hono/cors'
 import { zValidator } from '@hono/zod-validator'
-import { HTTPException } from 'hono/http-exception'
-import { ZodError } from 'zod'
 import { ConfigStorageService } from '@shared/config-storage-service'
 import { APP_URL, AWS_PREFIX } from '@shared/defines'
 import type { ConfigVersions } from '@shared/types'
-import * as probabilisticRevShare from './routes/probabilistic-revshare.js'
 import { OpenPaymentsService } from './utils/open-payments.js'
 import {
   PaymentQuoteSchema,
@@ -14,68 +9,10 @@ import {
   PaymentFinalizeSchema,
   WalletAddressParamSchema
 } from './schemas/payment.js'
-import { createHTTPException, serializeError } from './utils/utils.js'
+import { createHTTPException } from './utils/utils.js'
+import { app } from './app.js'
 
-export type Env = {
-  AWS_ACCESS_KEY_ID: string
-  AWS_SECRET_ACCESS_KEY: string
-  AWS_REGION: string
-  AWS_BUCKET_NAME: string
-  OP_WALLET_ADDRESS: string
-  OP_PRIVATE_KEY: string
-  OP_KEY_ID: string
-}
-
-const app = new Hono<{ Bindings: Env }>()
-
-app.use(
-  '*',
-  cors({
-    origin: '*',
-    allowMethods: ['GET', 'POST', 'OPTIONS'],
-    maxAge: 7200
-  })
-)
-
-app.onError((error, c) => {
-  if (error instanceof HTTPException) {
-    console.error(error)
-    const err = {
-      status: error.status,
-      statusText: error.res?.statusText,
-      message: error.message,
-      details: {
-        // @ts-expect-error if there's a cause, it should have a message
-        message: error.cause?.message
-      }
-    }
-    return c.json({ error: err }, error.status)
-  }
-
-  if (error instanceof ZodError) {
-    const err = {
-      message: 'Validation failed',
-      code: 'VALIDATION_ERROR',
-      details: {
-        issues: error.errors.map((err) => ({
-          path: err.path.join('.'),
-          message: err.message,
-          code: err.code
-        }))
-      }
-    }
-    return c.json({ error: err }, 400)
-  }
-
-  const serializedError = serializeError(error)
-  console.error('Unexpected error: ', serializedError)
-  const err = {
-    message: 'INTERNAL_ERROR',
-    ...serializedError
-  }
-
-  return c.json({ error: err }, 500)
-})
+import './routes/probabilistic-revshare.js'
 
 app.get(
   '/config/:wa/:version?',
@@ -169,21 +106,6 @@ app.post(
       return json(result)
     } catch (error) {
       throw createHTTPException(500, 'Payment finalization error: ', error)
-    }
-  }
-)
-
-app.get(
-  '/revshare/:payload',
-  zValidator('param', probabilisticRevShare.paramSchema),
-  async ({ req, json }) => {
-    const encodedPayload = req.param('payload')
-    try {
-      const result = await probabilisticRevShare.handler(encodedPayload)
-      return json(result)
-    } catch (error) {
-      if (error instanceof HTTPException) throw error
-      throw createHTTPException(500, 'Revenue share error', error)
     }
   }
 )
